polygon glass method
